AI_api/requirements.txt
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
flasgger==0.9.7.1
Flask==3.0.3
fonttools==4.53.0
fsspec==2024.6.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mistune==3.0.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
statsmodels==0.14.2
sympy==1.12.1
thop==0.1.1.post2209072238
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
ultralytics==8.0.196
urllib3==2.2.1
Werkzeug==3.0.3
